Improve repo structure
