+++ data/DISDRODB/Processed/DATA_SOURCE/CAMPAIGN_NAME/metadata/station_name_1.yml
@@ -1,47 +1,64 @@
-title: 'TITLE'
-description: ''
-source: ''
-history: ''
-conventions: ''
-campaign_name: 'CAMPAIGN_NAME'
-project_name: ''
-station_name: 'ID_station_1'
-station_name: ''
-station_number: '1'
-location: 'Cardada'
-country: 'Switzerland'
-continent: 'Europe'
-latitude: 46.2005117  
-longitude: 8.7916367
-altitude: 1671
+data_source: DATA_SOURCE
+campaign_name: CAMPAIGN_NAME
+station_name: station_name_1
+sensor_name: OTT_Parsivel
+reader: EPFL/LOCARNO_2018
+raw_data_format: raw
+raw_data_type: raw
+platform_type: fixed
 crs: WGS84
 proj4_string: +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
 EPSG: 4326
 latitude_unit: DegreesNorth
 longitude_unit: DegreesEast
 altitude_unit: MetersAboveSeaLevel
-sensor_name: OTT_Parsivel
+source: ''
+source_convention: ''
+source_processing_date: ''
+title: LOCARNO_2018
+description: ''
+history: ''
+keywords: ''
+conventions: ''
+project_name: ''
+comments: ''
+station_id: '60'
+location: Cimetta
+country: Switzerland
+continent: Europe
+latitude: 46.2005117
+longitude: 8.7916367
+altitude: 1671
+deployment_status: ''
+deployment mode: ''
+platform_protection: ''
+platform_orientation: ''
 sensor_long_name: OTT Hydromet Parsivel
-sensor_wavelegth: ''
+sensor_manufacturer: ''
+sensor_wavelength: ''
 sensor_serial_number: ''
-firmware_IOP: ''
-firmware_DSP: ''
+firmware_iop: ''
+firmware_dsp: ''
 firmware_version: ''
+sensor_beam_length: ''
 sensor_beam_width: ''
 sensor_nominal_width: ''
-temporal_resolution: ''
 measurement_interval: ''
+calibration_sensitivity: ''
+calibration_certification_date: ''
+calibration_certification_url: ''
 contributors: ''
 authors: ''
-institution: 'Laboratoire de Teledetection Environnementale -  Ecole Polytechnique Federale de Lausanne'
-reference: ''
-documentation: ''
-website: ''
-source_repository: ''
-doi: ''
+authors_url: ''
 contact: ''
 contact_information: http://lte.epfl.ch
-source_data_format: ''
-obs_type: ''
-level: ''
-disdrodb_id: '1'
+acknowledgement: ''
+references: ''
+documentation: ''
+website: ''
+institution: Laboratoire de Teledetection Environnementale -  Ecole Polytechnique
+  Federale de Lausanne
+source_repository: ''
+license: ''
+doi: ''
+summary: ''
