--- data/DISDRODB/Processed/DATA_SOURCE/CAMPAIGN_NAME/metadata/station_name_2.yml
@@ -1,47 +1,0 @@
-title: 'TITLE'
-description: ''
-source: ''
-history: ''
-conventions: ''
-campaign_name: 'CAMPAIGN_NAME'
-project_name: ''
-station_name: 'ID_station_2'
-station_name: ''
-station_number: '2'
-location: 'Locarno-Monti'
-country: 'Switzerland'
-continent: 'Europe'
-latitude: 46.1725900
-longitude: 8.7874267
-altitude: -9999
-crs: WGS84
-proj4_string: +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
-EPSG: 4326
-latitude_unit: DegreesNorth
-longitude_unit: DegreesEast
-altitude_unit: MetersAboveSeaLevel
-sensor_name: OTT_Parsivel
-sensor_long_name: OTT Hydromet Parsivel
-sensor_wavelegth: ''
-sensor_serial_number: ''
-firmware_IOP: ''
-firmware_DSP: ''
-firmware_version: ''
-sensor_beam_width: ''
-sensor_nominal_width: ''
-temporal_resolution: ''
-measurement_interval: ''
-contributors: ''
-authors: ''
-institution: 'Laboratoire de Teledetection Environnementale -  Ecole Polytechnique Federale de Lausanne'
-reference: ''
-documentation: ''
-website: ''
-source_repository: ''
-doi: ''
-contact: ''
-contact_information: http://lte.epfl.ch
-source_data_format: ''
-obs_type: ''
-level: ''
-disdrodb_id: '2'
