+++ docs/source/data.rst
@@ -3,14 +3,15 @@ Data
 =========================
 
 
-Access to the disdrometer measurement datasets is available through a separate Github repository : `disdrodb-data <https://github.com/ltelab/disdrodb-data>`__.
+Access to the disdrometer measurements and metadata is available through a separate Github repository : `disdrodb-data <https://github.com/ltelab/disdrodb-data>`__.
 
 
-This GitHub repository includes the following information:
+The `disdrodb-data` GitHub repository includes:
 
-- Stations measurements in the form of a JSON file with links to the downloadable data (not included in the repository due to GitHub's file size limitation).
-- Stations metadata in YAML files.
-- Stations issues in YAML files.
-- Python code to download the datasets using the URLs provided in the JSON files.
+- The stations measurements in the form of a JSON file with URLs links to the downloadable data (not included in the repository due to GitHub's file size limitation).
+- The stations metadata in YAML file format.
+- The stations issue in YAML file format.
+- Some python code to download the datasets using the URLs provided in the JSON files.
+- Some python code to curate the metadata archive.
