+++ docs/source/installation.rst
@@ -34,9 +34,9 @@ Installation for contributors
 ================================
 
 
-The latest disdrodb stable version is available on the Github repository `disdrodb <https://github.com/ltelab/disdrodb>`_.
+The latest disdrodb stable version is available on the GitHub repository `disdrodb <https://github.com/ltelab/disdrodb>`_.
 
-Clone the repository from github
+Clone the repository from GitHub
 .........................................
 
 According to the `contributors guidelines <contributors_guidelines>`__, you should first create a fork into your personal GitHub account.
@@ -56,7 +56,7 @@ According to the `contributors guidelines <contributors_guidelines>`__, you shou
 Install the DISDRODB package dependencies
 ............................................
 
-You can use either conda or pip : 
+You can use either pip or conda: 
 
 
 * **Pip**
@@ -86,14 +86,6 @@ To install the project in editable mode :
 
 
 
-
-
-
-
-
-
-
-
 Install pre-commit code quality checking
 ..............................................
 
@@ -110,17 +102,16 @@ root:
 The pre-commit hooks are scripts executed automatically in every commit
 to identify simple code quality issues. When an issue is identified
 (the pre-commit script exits with non-zero status), the hook aborts the
-commit and prints the error. Currently, DISDRODB only tests that the
-code to be committed complies with black’s format style. 
+commit and prints the error. Currently, DISDRODB tests that the
+code to be committed complies with `black’s  <https://github.com/psf/black>`__ format style
+and the `ruff <https://github.com/charliermarsh/ruff>`__ linter.
 
-In case that the commit is aborted, you only need to run black agains you code.
-This can be done by running ``black .`` or ``pre-commit run --all-files``. The latter is recommended since it
-indicates if the commit contained any formatting errors (that are automatically corrected).
+In case that the commit is aborted, you only need to run `black`and `ruff` through your code.
+This can be done by running ``black .`` and ``ruff check .`` or alternatively with ``pre-commit run --all-files``. 
+The latter is recommended since it indicates if the commit contained any formatting errors (that are automatically corrected).
 
 .. note::
 	To maintain consitency, we use Black version `22.8.0` (as defined into `.pre-commit-config.yaml`). Make sure to stick to version.  
-
-
 
 
 
@@ -153,7 +144,7 @@ To set up a virtual environment, follow these steps :
 
 
 
-* **With Conda:**
+* **With conda:**
 
 	* Create the `disdrodb-dev` (or anay other name) conda environment:
 
@@ -167,6 +158,18 @@ To set up a virtual environment, follow these steps :
 
 			conda activate disdrodb-dev
 
-		
+Run DISDRODB on Jupyter Notebooks
+==================================	
+
+If you want to run disdrodb on a `Jupyter Notebook <https://jupyter.org/>`__, 
+you have to take care to set up the IPython kernel environment where disdrodb is installed.
+
+For example, if your conda/virtual environment is named `disdrodb-dev`, run:
+
+.. code-block:: bash 
+
+    python -m ipykernel install --user --name=disdrodb-dev
+
+When you will use the Jupyter Notebook, by clicking on `Kernel` and then `Change Kernel`, you will be able to select the `disdrodb-dev` kernel.
