+++ docs/source/metadata.rst
